trigger delete action only if at least one row is selected
git-svn-id: file:///www/metaways/svn.tine20.org/repo/trunk@700 8d11cdf3-9d2f-0410-bbb5-e098f187f06b

+++ tine20/Admin/js/Admin.js
@@ -1069,7 +1069,7 @@ Egw.Admin.Accounts.Main = function() {
 	        });
 	        
 	        grid_applications.on('keydown', function(e){
-                 if(e.getKey() == e.DELETE){
+                 if(e.getKey() == e.DELETE && Ext.getCmp('AdminAccountsGrid').getSelectionModel().getCount() > 0){
                      this.deleteButtonHandler();
                  }
             }, this);
